Tratando exceções para mostrar o código de resposta personalizado

+++ src/main/java/gabriel/hb/MyLifeBackend/services/UserService.java
@@ -4,10 +4,14 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import gabriel.hb.MyLifeBackend.entities.User;
 import gabriel.hb.MyLifeBackend.repositories.UserRepository;
+import gabriel.hb.MyLifeBackend.services.exceptions.DatabaseException;
+import gabriel.hb.MyLifeBackend.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service // Registra a classe como um componente/service do spring e vai poder ser injetado no UserResource
 public class UserService {
@@ -28,15 +32,28 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-    public void delete (long id) {
-        repository.deleteById(id);
-    }
+	public void delete(Long id) {
+	    try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id); // Lança uma exceção através do 'ResourceExceptionHandler', que captura as excecões que ocorrem		
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
+	} 
 	
-    public User update(long id, User obj) {
-        User entity = repository.getReferenceById(id); // Deixa um obj monitorado pelo JPA, não realizando operação com o banco igual o findById
-        updateData(entity, obj);
-        return repository.save(entity);
-    }
+	public User update(long id, User obj) {
+		try {
+			User entity = repository.getReferenceById(id); // Deixa um obj monitorado pelo JPA, não realizando operação com o banco igual o findById
+			updateData(entity, obj);
+			return repository.save(entity);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
+	}
 
 	private void updateData(User entity, User obj) {
 		entity.setUsername(obj.getUsername());
